Redacted MEGA
Removed MEGA from cloud storage providers due to the seizure of the site and all assets by the New Zealand government in 2015. Additionally, due to new developments in Australian & New Zealand privacy laws, it is highly recommended to avoid any and all services hosted in or operated in either country.